Abstract useManyThrough forms functionality to a resuable hook
