+++ packages/react/src/auto/shadcn/inputs/relationships/ShadcnAutoHasManyThroughForm.tsx
@@ -1,12 +1,10 @@
 import { ChevronsUpDown, PlusIcon, X } from "lucide-react";
-import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { useFormContext } from "../../../../useActionForm.js";
+import React, { useCallback, useState } from "react";
+import { useHasManyThroughForm } from "../../../../useHasManyThroughForm.js";
 import { debounce } from "../../../../utils.js";
-import { extractPathsFromChildren } from "../../../AutoForm.js";
 import { autoRelationshipForm } from "../../../AutoInput.js";
-import { RelationshipContext, useAutoRelationship, useRelationshipContext } from "../../../hooks/useAutoRelationship.js";
-import { useHasManyThroughController } from "../../../hooks/useHasManyThroughController.js";
-import { getRecordAsOption, useOptionLabelForField } from "../../../hooks/useRelatedModel.js";
+import { RelationshipContext } from "../../../hooks/useAutoRelationship.js";
+import { getRecordAsOption } from "../../../hooks/useRelatedModel.js";
 import type { Option, OptionLabel } from "../../../interfaces/AutoRelationshipInputProps.js";
 import type { ShadcnElements } from "../../elements.js";
 import { makeShadcnRenderOptionLabel } from "../../utils.js";
@@ -89,64 +87,29 @@ export const makeShadcnAutoHasManyThroughForm = ({
     secondaryLabel?: OptionLabel;
     tertiaryLabel?: OptionLabel;
   }) {
-    const { field, children } = props;
-    const { metadata } = useAutoRelationship({ field });
-    const { setValue } = useFormContext();
     const [open, setOpen] = useState(false);
-    const childPaths = children && extractPathsFromChildren(children);
-    const hasChildForm = childPaths && childPaths.length > 0;
-
-    const { fieldArrayPath, fieldArray, records, relatedModelOptions, inverseRelatedModelField, joinModelField, joinModelApiIdentifier } =
-      useHasManyThroughController(props);
-
-    const { fields, append, remove } = fieldArray;
-    const relationshipContext = useRelationshipContext();
-    const pathPrefix = relationshipContext?.transformPath ? relationshipContext.transformPath(props.field) : props.field;
-    const metaDataPathPrefix = relationshipContext?.transformMetadataPath
-      ? relationshipContext.transformMetadataPath(props.field)
-      : pathPrefix;
-
-    useEffect(() => {
-      for (const [index, field] of fields.entries()) {
-        if (
-          inverseRelatedModelField &&
-          (field as any)[inverseRelatedModelField] &&
-          (field as any)[inverseRelatedModelField].id &&
-          !("_link" in field)
-        ) {
-          // use setValue so that we don't trigger a re-render
-          setValue(`${fieldArrayPath}.${index}.${inverseRelatedModelField}._link`, (field as any)[inverseRelatedModelField].id);
-        }
-      }
-    }, [setValue, inverseRelatedModelField, fields, fieldArrayPath]);
-
-    const listboxId = `HasManyThroughListboxInPopover-${field}`;
-
-    if (metadata?.configuration.__typename !== "GadgetHasManyThroughConfig") {
-      throw new Error("ShadcnAutoHasManyThroughForm can only be used for HasManyThrough fields");
-    }
-
     const {
+      field,
+      append,
+      remove,
+      joinRecords,
+      metadata,
+      primaryLabel,
+      hasChildForm,
+      listboxId,
+      pathPrefix,
+      metaDataPathPrefix,
+      siblingModelName,
+      siblingRecordsLoading,
+      siblingRecords,
+      siblingPagination,
       search,
-      searchFilterOptions: siblingModelOptions,
-      relatedModel: { fetching: siblingRecordsLoading, records: siblingRecords },
-      pagination: siblingPagination,
-    } = relatedModelOptions;
-    const siblingModelName = metadata.name ?? "Unknown";
-
-    const joinRecords = useMemo(() => {
-      return fields.flatMap((field, idx): [string, number, Record<string, any>][] => {
-        const record = records[idx];
-
-        if (!record) {
-          return [];
-        }
-
-        return [[field._fieldArrayKey, idx, record]];
-      });
-    }, [fields, records]);
-
-    const primaryLabel = useOptionLabelForField(field, props.primaryLabel);
+      joinModelField,
+      joinModelApiIdentifier,
+      siblingModelOptions,
+      inverseRelatedModelField,
+      fieldArray,
+    } = useHasManyThroughForm(props);
 
     const handleScrolledToBottom = useCallback(
       debounce(() => {
